Раскидывает код на компоненты
