+++ src/const.js
@@ -1,0 +1,26 @@
+export const MONTHS = [
+  `Jan`,
+  `Feb`,
+  `Mar`,
+  `Apr`,
+  `May`,
+  `Jun`,
+  `July`,
+  `Aug`,
+  `Sept`,
+  `Oct`,
+  `Nov`,
+  `Dec`
+];
+
+export const DATE_VALUE = {
+  day: 86400000,
+  hour: 3600000,
+  minutes: 6000
+};
+
+export const SPREADS_DAY = [
+  86401900,
+  17283200,
+  259209150
+];
