+++ src/mock/trip.js
@@ -1,31 +1,11 @@
 import {getRandomInteger} from "../utils";
+import {parseDateTimeFormat} from "../utils";
+import {parseDateDayFormat} from "../utils";
+import {getDateDiff} from "../utils";
+import {SPREADS_DAY} from "../const";
 
-const SPREADS_DAY = [
-  86401900,
-  17283200,
-  259209150
-];
-
-const DAY = 86400000;
-const HOUR = 3600000;
-const MINUTES = 6000;
 const MIN_TIME_INTERVAL = 1200000;
 const MAX_TIME_INTERVAL = 10800000;
-
-const MONTHS = [
-  `Jan`,
-  `Feb`,
-  `Mar`,
-  `Apr`,
-  `May`,
-  `Jun`,
-  `July`,
-  `Aug`,
-  `Sept`,
-  `Oct`,
-  `Nov`,
-  `Dec`
-];
 
 const pointTypes = [
   `Taxi`,
@@ -88,16 +68,6 @@ const getOffers = () => {
   }
 };
 
-export const parseDateDayFormat = (date) => {
-  return MONTHS[date.getMonth()] + ` ` + date.getDate();
-};
-
-export const parseDateTimeFormat = (date) => {
-  let currentDate = new Date(date);
-  currentDate = currentDate.toTimeString().split(`:`);
-  return currentDate[0] + `-` + currentDate[1];
-};
-
 const generateStartDate = () => {
   const currentDate = Date.parse(new Date());
   return new Date(SPREADS_DAY[getRandomInteger(0, SPREADS_DAY.length - 1)] + currentDate);
@@ -106,31 +76,6 @@ const generateStartDate = () => {
 const generateFinishDate = (startDate) => {
   const currentDate = Date.parse(startDate);
   return new Date(getRandomInteger(MIN_TIME_INTERVAL, MAX_TIME_INTERVAL) + currentDate);
-};
-
-const getDateDiff = (dateStart, dateFinish) => {
-  const start = Date.parse(dateStart);
-  const finish = Date.parse(dateFinish);
-  let currentDate = finish - start;
-  let day = 0;
-  let hour = 0;
-  let minutes = 0;
-
-  if (currentDate > DAY) {
-    day = Math.floor(currentDate / 1000 / 60 / 60 / 24).toFixed(0);
-    currentDate = currentDate - (day * DAY);
-  }
-
-  if (currentDate > HOUR) {
-    hour = Math.floor(currentDate / 1000 / 60 / 60).toFixed(0);
-    currentDate = currentDate - (hour * HOUR);
-  }
-
-  if (currentDate > MINUTES) {
-    minutes = Math.floor(currentDate / 1000 / 60).toFixed(0);
-  }
-
-  return {day, hour, minutes};
 };
 
 export const renderPointTripItem = () => {
